requirements.txt
psycopg2-binary==2.9.3
anyio==3.5.0
arrow==1.2.1
asgiref==3.4.1
asn1crypto==1.4.0
astroid==2.9.3
bidict==0.21.4
blessed==1.19.0
Brotli==1.0.9
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.47.0
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-picklefield==3.0.1
django-ses==2.3.1
django-q==1.3.9
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.4
httpx==0.21.3
idna==3.3
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
notion-client==0.8.0
oauthlib==3.1.1
pycparser==2.21
PyJWT==2.3.0
pylint==2.12.2
pyOpenSSL==21.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.3.1
python-socketio==5.5.1
python3-openid==3.2.0
pre-commit==2.16.0
pytz==2021.3
redis==4.1.0
requests==2.27.1
requests-oauthlib==1.3.0
rfc3986==2.0.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
typed-ast==1.5.1
typing-extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.2
whitenoise[brotli]==5.3.0
wrapt==1.13.3
xlrd==2.0.1
